# docs/swagger.yaml
basePath: /
definitions:
  api_teklifYonetimi_internal_api_response.APIResponse:
    properties:
      code:
        type: string
      data: {}
      message:
        type: string
      meta: {}
      success:
        type: boolean
    type: object
...
  api_teklifYonetimi_internal_dto.CreateUserRequest:
    properties:
      company_id:
        type: integer
      email:
        type: string
      name:
        type: string
      password:
        minLength: 6
        type: string
      role:
        type: string
    required:
    - email
    - name
    - password
    type: object
...
  api_teklifYonetimi_internal_dto.UpdateUserRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        minLength: 6
        type: string
      role:
        type: string
    type: object
host: localhost:8082
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Teklif Yönetim Sistemi API Dokümantasyonu
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Teklif Yönetimi API
  version: "1.0"
paths:
...
  /users:
    get:
      consumes:
      - application/json
      description: Tüm kullanıcıları listeler
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api_teklifYonetimi_internal_api_response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api_teklifYonetimi_internal_api_response.APIResponse'
      security:
      - BearerAuth: []
      summary: Kullanıcıları Listele
      tags:
      - Users
    post:
      consumes:
      - application/json
      description: Yeni bir kullanıcı kaydı oluşturur
      parameters:
      - description: Kullanıcı Bilgileri
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/api_teklifYonetimi_internal_dto.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/api_teklifYonetimi_internal_api_response.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api_teklifYonetimi_internal_api_response.APIResponse'
      security:
      - BearerAuth: []
      summary: Yeni Kullanıcı Oluştur
      tags:
      - Users
  /users/{id}:
    delete:
      consumes:
      - application/json
      description: Kullanıcıyı siler
      parameters:
      - description: Kullanıcı ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api_teklifYonetimi_internal_api_response.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api_teklifYonetimi_internal_api_response.APIResponse'
      security:
      - BearerAuth: []
      summary: Kullanıcı Sil
      tags:
      - Users
    get:
      consumes:
      - application/json
      description: ID'ye göre kullanıcı detayını getirir
      parameters:
      - description: Kullanıcı ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api_teklifYonetimi_internal_api_response.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api_teklifYonetimi_internal_api_response.APIResponse'
      security:
      - BearerAuth: []
      summary: Kullanıcı Detayı
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: Kullanıcı bilgilerini günceller
      parameters:
      - description: Kullanıcı ID
        in: path
        name: id
        required: true
        type: integer
      - description: Güncelleme Bilgileri
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/api_teklifYonetimi_internal_dto.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api_teklifYonetimi_internal_api_response.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api_teklifYonetimi_internal_api_response.APIResponse'
      security:
      - BearerAuth: []
      summary: Kullanıcı Güncelle
      tags:
      - Users
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"